v1.3 funciona

+++ lib/execution_monitoring/domain/entities/location.dart
@@ -1,7 +1,7 @@
 class Location {
   final int id;
-  final int lat;
-  final int lng;
+  final double lat;
+  final double lng;
 
   Location({required this.id, required this.lat, required this.lng});
 }
